bau: Reset pay-java-commons repo
Given that PP-3842 is essentially the same story as PP-3615 - they are stories
to create and run provider pact tests - and custom pact classes have been
created to make pact testing work for us, it seemed sensible to extract those
pact classes into a common project for use by all projects that run pact
provider tests. After some digging it turns out this repo (pay-java-commons) is
meant to be the repo for common classes. However it is not currently being used
so this commit takes the opportunity to reset it by deleting unused stuff
before the pact testing utility classes are put in here.

@oswaldquek