bau: PactProviderRule for junit consumer tests
This commit commits a PactProviderRule which is almost a direct copy of
https://github.com/DiUS/pact-jvm/blob/master/pact-jvm-consumer-junit/src/main/java/au/com/dius/pact/consumer/BaseProviderRule.java.
The rationale for doing this is that we want each unit test (see
PactProviderRuleTest for an example) to define its own pact interactions via
the @Pacts annotation. The BaseProviderRule provided by github.com/DiUS allows
us to do this by using the `fragment` property of the `PactVerification`
annotation (for e.g. `@PactVerification(value = "provider", fragment =
"createFragment2")`); however we're not using `fragment`s as they are created
using java code, rather we're reading in the pact files direct via the @Pacts
annotation. As a result we've had to override the `apply` and `getPacts` function.
However many other methods are not overridable so we've mostly copied the
PactProviderRule.

@oswaldquek