PP-3845: Add ability to not publish pacts
PP-3845 is about checking pact compatibility, e.g. can I deploy service A to
staging given there is service B in staging as well. The issue is sometimes
consumer pact tests are written in development and by default are published to
the pact broker - at this stage no provider tests have been written. This
causes pact compatibility check to fail as the pact broker now has knowledge of
the new pact.

This commit gives the option to not publish a pact by deleting it from the
target/pacts directory, which the pact:publish maven plugin looks into to
publish pacts to the broker.

It would be nice to specify what pacts not to publish in the pact:publish
plugin, an issue has been raised here
https://github.com/DiUS/pact-jvm/issues/711.

@oswaldquek