bau: Load pacts per test
Before this commit, the PactProviderRule loads ALL the pacts for ALL tests,
even if only one test is being run (from the IDE). This was causing problems
for some developers because they defined multiple pacts to the same provider,
each one being annotated on a single unit test. But because PactProviderRule
was loading all the pacts in the test class, the pacts would get overwritten.
The new test in PactProviderRuleTest in this commit demonstrates this: there
are two tests - getAmexPrice and getVisaPrice - calling the same provider.
Prior to this commit, one of these would fail randomly. This no longer happens
because PactProviderRule only loads the pact for that unit test.

@oswaldquek