PP-3845: Move pact publishing test to its own class
Pact files get written to target/pacts directory after the unit test itself is
evaluated, so the assertions have to be written in the @AfterClass rather than
in the @Test method. Moving the pact publishing test to its own test as having
assertions in the @AfterClass can cause confusion if placed in
PactProviderRuleTest.

@oswaldquek