PP-2495: Initialize AWS Xray within a static method
It turns out using Bundles does not work. I am not completely sure why but
looking at https://www.dropwizard.io/1.0.5/docs/manual/internals.html it is
possible this is either due to:

1. the Cli not being instantiated before the `bundle.initialize(bootstrap)` is
run
2. the environment is only created during the `cli.run` phase so the xray stuff
needs to be instantiated after the environment is created, not before when
`bundle.initialize(bootstrap)` is run during the `initialize(bootstrap)` phase.

With this commit, dropwizard apps simply need to call Xray.init in the
Application's run method.

@oswaldquek