PP-4474: Import postgres container handling code from adminusers
We have copies of this code in all our projects for bringing up and tearing
down a postgres container in order to run tests. Bring it into our central
testing library and then share it back out.

This has been refactored slightly to provide a default constructor for
PostgresContainer which takes no parameters.