PP-4595 ISO instant date/time formatter with millisecond precision
We use DateTimeFormatter.ISO_INSTANT.withZone(UTC) to format a lot
of the timestamps we put in API responses, which produces strings
like "2015-10-21T07:28:00.123Z". The fractional part, if present,
will have either three, six or nine digits, depending on the
precision available.

When one asks Java 8 for the current time, it always returns the
time with millisecond precision, even if the underlying system
clock can do better. Java 9 and above don’t have this limitation
and will return higher accuracy if the system clock offers it (see
https://bugs.openjdk.java.net/browse/JDK-8068730 for details). On
Linux and macOS, this usually means microsecond precision.

Therefore, when DateTimeFormatter.ISO_INSTANT.withZone(UTC) is
used to format the current time on Java 8, the resulting string has
three digits after the decimal point. When it’s used on Java 9 on
Linux or macOS, it will have six digits. This means that the
timestamps we put into API responses — many of which are the
current time or a stored past current time — may vary depending
on the Java version.

Eliminate this difference by introducing a DateTimeFormatter that
always uses millisecond precision. If this DateTimeFormatter is
used throughout the system, then there will be no change to the
timestamps in our API responses when we upgrade to Java 11.

Note that this new DateTimeFormatter has one slight behaviour
change: if DateTimeFormatter.ISO_INSTANT.withZone(UTC) is asked
to format a time that’s exactly on the second, it will not include
a fractional component (e.g. "2015-10-21T07:28:00Z"). The new
DateTimeFormatter always includes a fractional component (e.g.
"2015-10-21T07:28:00.000Z"). This actually makes things more
consistent, so we’re happy with this change.