PP-4616: Add a convenience constructor to ApplicationStartupDependentResourceChecker
It's only the tests which need to pass in a custom 'waiter' to this class, so
provide a constructor that supplies a sensible default implementation for the
real consumers to use, making their lives easier.

Stop using InjectMocks in the test because it's too magic and selects the wrong
constructor.