PP-4790 Get metadata from database login
- Most of the metadata can be accessed independent of a connected session.
This is done using the DatabaseLogin object. This metadata can be skipped if
not available so we will populate only what we can find.
- We decided to avoid querying if the connection is null, as we observed that
after a database error, all subsequent calls to the profiler were using a
`disconnected` database session. We don't need to use a connection, no matter
what the status, so we will not use it at all.
- Added DropWizard testing to write a small unit test to check that the
properties are set