PP-4886: Serialise empty metadata string to null object
It's been agreed with @danworth and @alexbishop1 we should be lenient and
accept create payment requets where metadata is empty:

```
"metadata": {}
```

If this happens we want to treat the payment as if it had no metadata.
Therefore this needs to be serialized to null which means the payment will have
no jsonb metadata in the database which is what we want.