PP-5318 Abstract type for value types that wrap strings
Add an abstract class that wraps a string. This can make the
implementation of our burgeoning collection of specific value
types (types that represent particular IDs etc.) simpler because
they don’t each need to implement equals(…) and the like (which
can be surprisingly hard to get right sometimes!).

The wrapped string must not be null. To be considered equal, two
objects must have both the same runtime type and the same string
value: instances of different subtypes that wrap the same string
are not equal.

Use it like this:

public class ExternalPaymentId extends WrappedStringValue {

    private ExternalPaymentId(String id) {
        super(id);
    }

    public static ExternalPaymentId valueOf(String id) {
        return new ExternalPaymentId(id);
    }

}