PP-5542 Initial structured logging keys for Java applications
An interface of string constants to be used as structured logging
keys in an attempt to encourage consistency.

Generally, key names are quite broad, for example “provider_id”,
which could be a provider’s ID for a payment, mandate, refund etc.
However, the data recorded could compose multiple keys to aid
disambiguation. For instance, one could have a log line with
properties like this:

{
  "provider_id": "foo",
  "provider": "sandbox",
  "payment-type": "card",
  "resource_type": "payment"
}

Here, it’s clear that the “providerId” refers to a sandbox card
payment. Someone searching the logs who has a provider ID but
doesn’t know what it refers to could just search for the
“provider_id”, while someone who knows they want log lines related
to a sandbox card payment could add the extra properties into their
search.

with @oswaldquek