PP-5670: Use multivalued map
CreateTestSuite was originally created with a one-to-one mapping of consumer to
test class in mind. See
https://github.com/alphagov/pay-connector/blob/master/src/test/java/uk/gov/pay/connector/pact/ContractTestSuite.java
as an example where each consumer has one JUnit test class.

Pay-ledger breaks this as the pacts between connector and itself cover a span
of test class.
https://github.com/alphagov/pay-ledger/blob/master/src/test/java/uk/gov/pay/ledger/pact/ContractTestSuite.java
has a number of classes for the connector->ledger interaction:
PaymentDetailsEnteredEventQueueContractTest,
PaymentCreatedEventQueueContractTest, CaptureConfirmedEventQueueContractTest,
RefundCreatedByUserEventQueueContractTest,
RefundSubmittedEventQueueContractTest etc.

This PR changes CreateTestSuite so it accepts a multi valued map which allows a
key ("connector" in this case) to map to multiple test classes.