PP-7242 Introduce a CardExpiryDate type for, er, card expiry dates
Introduce a CardExpiryDate type to model card expiry dates. This
is essentially a wrapper around a 2-digit month and a 2-digit year,
which are the hard-working aspects of a card expiry date.

The type can be constructed from a string in MM/yy format and has
a toString() method that converts it back to this format.

In addition, it has convenience methods to get the month as a
2-digit string and the year as a 2-digit string or a 4-digit
string (for the latter, the 2-digit string is prefixed with "20",
ingraining but also centralising and encapsulating the assumption
that already pervades our system).

The type also offers a method to produce to an equivalent YearMonth
(again, assuming the year begins with "20").

As a bonus, the type exposes a regex pattern that can be used to
validate a string is in the required MM/yy format (we currently
have at least 2 such regexen in our system and, no, they’re not
identical in behaviour).

The type follows the convention of having a static valueOf(String)
method that Jackson can automatically use when deserialising a JSON
payload containing an expiry date in MM/yy format into a Java
object. The fact that toString() is used to produce the MM/yy form
means it also plays nicely with Jackson’s ToStringSerializer when
going from a Java object to JSON.

The new CardExpiryDateConverter allows for converting from a
CardExpiryDate to a string and vice-versa to make it easy to use
CardExpiryDate in entities that are persisted to the database.