PP-7242 Rename some CardExpiryDate getters
In CardExpiryDate, rename some getters:

• get2DigitMonth() → getTwoDigitMonth()
• get2DigitYear()  → getTwoDigitYear()
• get4DigitYear()  → getFourDigitYear()

We are going to use these methods a lot in Freemarker templates. If
you have an object foo that has a method getWhatever(), Freemarker
lets you do the Java Beans-esque ${foo.whatever} but this does not
work if the first character after “get” in the method name is a
digit.

This will really annoy me and it’s easy to fix earlier on so this
is me doing that.