PP-7205 Add/modify logging keys for request logging
Modify the HTTP_STATUS logging key to be "status_code" as this is what
we use for request logging provided by dropwizard, and is also what we
use in our Node apps.

Have confirmed that we don't use this key anywhere in our applications
currently by doing a search in alphagov in github.

Add additional keys for "url", "method" and "response_time" so these can
be used in logging for requests that we do in addition to the dropwizard
logging. These other log lines contain additional information such as
resource ids involved in the request. Modify the dropwizard logging
template to use these keys to ensure we are consistent.