PP-7205 Log request id for dropwizard request logs
Make our implementation of the dropwizard AccessJsonLayoutBaseFactory
even more custom by using our own implementation of AbstractJsonLayout
to build the log line for the request.

This is required so that we can read the "X-Request-Id" header and log
it as a key in the root object. If we used the dropwizard
AccessJsonLayout class, it is only possibly to log this as a field
in a "headers" object.