PP-7104 Log the forwarded for IP address
We were logging the `remote_address` for all Dropwizard requests, but
this was logging the address of the NGINX proxy.

Use the "X-Forwarded-For" header which contains a list of the client IP
address and the IP addressed of the proxies the request passes through.
Extract just the client IP address to log.

Add a logging key for "remote_address", as we also want to log using
this key in public api when an invalid API key is used.