PP-7590 Tweak ApiResponseDateTimeFormatter for Instants
Update Javadoc for ApiResponseDateTimeFormatter to clarify it’s
not just for ZonedDateTimes, moving the Javadoc to the
DateTimeFormatter field rather than its containing class.

Add some tests to demonstrate that ApiResponseDateTimeFormatter
works for Instants.

Do not set a time zone on the DateTimeFormatterBuilder because it’s
never actually used (appending an ISO instant to a
DateTimeFormatterBuilder always adds it using UTC). This does not
change how dates and times are formatted.