PP-7590 JPA converter to convert between Instant and LocalDateTime
Add a JPA AttributeConverter that converts between an Instant
entity attribute and a LocalDateTime (assumed to be in UTC) for
storing in a database column and vice versa.

Per JDBC 4.2, LocalDateTime maps to database columns of the type
TIMESTAMP WITHOUT TIME ZONE.

This class is intended to be a replacement for
UTCDateTimeConverter (found in connector and other services)
except that it works with Instants rather than ZonedDateTimes.

Contrary to its name, UTCDateTimeConverter’s actually stores
dates and times in the database using the default time zone.
The input ZonedDateTime gets converted to a Timestamp, which is
effectively just a count of nanoseconds since the epoch. The
database driver then has to store this Timestamp in a TIMESTAMP
WITHOUT TIME ZONE database column. It does this by converting the
Timestamp into the equivalent local date and time in the default
time zone. This is fine if the default time zone is UTC. However,
if the default time zone is, say, Europe/London and it’s British
Summer Time then all times will be shifted forward an hour and
10:00 will become 11:00.

This class can also replace connector’s LocalDateTimeConverter
(again, except that it works with Instants rather than
ZonedDateTimes), which was created to work like
UTCDateTimeConverter but without the default time zone issue.

Since both LocalDateTime and TIMESTAMP WITHOUT TIME ZONE are just
dates and times with no time zone information, the default time
zone is not required to convert between one and the other.[1]

Since our production Java servers have always used UTC as their
default time zone, there will be no change to how existing
dates and times in the database are interpreted. There may be
changes in environments that use a different default time zone,
such as developer machines, but such environments will only have
unimportant test data.

If we ever change our database columns to be of type TIMESTAMP
WITH TIME ZONE (which some would recommend), we would have to
use a converter that works with OffsetDateTimes instead of
LocalDateTimes.

[1] Unfortunately, this is not actually true. The PostgreSQL JDBC
    driver does actually use the default time zone when storing a
    LocalDateTime in a TIMESTAMP WITHOUT TIME ZONE database
    column.[2] However, this only causes the date and time to be
    changed if the LocalDateTime that doesn’t exist in the default
    time zone (e.g. the hour that’s skipped when the clocks go
    forward). Fortunately, this doesn’t happen very often.

[2] See https://github.com/pgjdbc/pgjdbc/issues/1390#issuecomment-698631932
    for my attempt to persuade the developers to make it not do
    this.