PP-7938: Pin action-maven-publish github action to a specific commit
Using github actions comes with security risks. A hacker could raise a PR on a
repo that injects github actions and tries to get [our maven gpg] secrets for
example. In order mitigate this we considered a few options. The following ones
were discounted:

* Depending on a major version. This was discounted as it would mean us blindly
  accepting a minor bump for it.
* Depending on a minor version. We don't have confidence that these versions
  are easily overwritable.
* Writing a script to publish and release our code to maven repo ourselves.
  This is the best option but would take another day.

The team have agreed that depending on a commit sha is a good compromise and
these can't be changed. Any force pushed changes to the github actions's
history would cause the action to not be able to run (which is a good thing).